extract EnigmaProjectImpl
